fix: error saving visa request and refactor org search field
