+++ apps/lfx-one/src/app/modules/events/my-events-dashboard/components/visa-request-apply-form/visa-request-apply-form.component.ts
@@ -1,32 +1,27 @@
 // Copyright The Linux Foundation and each contributor to LFX.
 // SPDX-License-Identifier: MIT
 
-import { ChangeDetectionStrategy, Component, computed, inject, output, Signal, signal } from '@angular/core';
-import { takeUntilDestroyed, toObservable, toSignal } from '@angular/core/rxjs-interop';
-import { FormControl, FormGroup, NonNullableFormBuilder, ReactiveFormsModule, Validators } from '@angular/forms';
-import { EventsService } from '@app/shared/services/events.service';
+import { ChangeDetectionStrategy, Component, inject, output } from '@angular/core';
+import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
+import { NonNullableFormBuilder, ReactiveFormsModule, Validators } from '@angular/forms';
 import { UserService } from '@app/shared/services/user.service';
-import { AutocompleteComponent } from '@components/autocomplete/autocomplete.component';
 import { CalendarComponent } from '@components/calendar/calendar.component';
 import { InputTextComponent } from '@components/input-text/input-text.component';
 import { SelectComponent } from '@components/select/select.component';
 import { TextareaComponent } from '@components/textarea/textarea.component';
 import { ACCOMMODATION_PAID_BY_OPTIONS, ATTENDEE_TYPE_OPTIONS, COUNTRIES } from '@lfx-one/shared/constants';
-import { AttendeeAccommodationPaidBy, AttendeeType, OrgSearchResult, VisaRequestApplicantInfo } from '@lfx-one/shared/interfaces';
-import { MessageService } from 'primeng/api';
-import { AutoCompleteCompleteEvent, AutoCompleteSelectEvent } from 'primeng/autocomplete';
-import { catchError, debounceTime, map, of, startWith, switchMap } from 'rxjs';
+import { AttendeeAccommodationPaidBy, AttendeeType, VisaRequestApplicantInfo } from '@lfx-one/shared/interfaces';
+import { startWith } from 'rxjs';
+import { OrgSearchFieldComponent } from '../org-search-field/org-search-field.component';
 
 @Component({
   selector: 'lfx-visa-request-apply-form',
-  imports: [ReactiveFormsModule, InputTextComponent, SelectComponent, TextareaComponent, CalendarComponent, AutocompleteComponent],
+  imports: [ReactiveFormsModule, InputTextComponent, SelectComponent, TextareaComponent, CalendarComponent, OrgSearchFieldComponent],
   templateUrl: './visa-request-apply-form.component.html',
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class VisaRequestApplyFormComponent {
   private readonly userService = inject(UserService);
-  private readonly eventsService = inject(EventsService);
-  private readonly messageService = inject(MessageService);
   private readonly fb = inject(NonNullableFormBuilder);
 
   public readonly formValidityChange = output<boolean>();
@@ -48,19 +43,9 @@ export class VisaRequestApplyFormComponent {
     attendeeAccommodationPaidBy: ['' as AttendeeAccommodationPaidBy, Validators.required],
   });
 
-  /** Separate form used only for the autocomplete binding — keeps the main form controls clean */
-  public readonly orgForm = new FormGroup({ company: new FormControl<OrgSearchResult | null>(null) });
-
   public readonly countryOptions = [...COUNTRIES];
   public readonly attendeeTypeOptions = ATTENDEE_TYPE_OPTIONS;
   public readonly accommodationOptions = ACCOMMODATION_PAID_BY_OPTIONS;
-
-  protected readonly orgSearchQuery = signal('');
-  protected readonly orgSuggestions: Signal<OrgSearchResult[]> = this.initOrgSuggestions();
-
-  protected readonly isCompanyInvalid = computed(
-    () => this.form.get('company')?.invalid && (this.form.get('company')?.touched || this.orgForm.get('company')?.touched || this.orgForm.get('company')?.dirty)
-  );
 
   public constructor() {
     this.form.get('email')?.disable();
@@ -75,59 +60,12 @@ export class VisaRequestApplyFormComponent {
     }
 
     this.form.statusChanges.pipe(startWith(this.form.status), takeUntilDestroyed()).subscribe(() => {
-      this.formValidityChange.emit(this.form.valid && !this.isCompanyInvalid());
+      this.formValidityChange.emit(this.form.valid);
     });
 
     this.form.valueChanges.pipe(startWith(this.form.getRawValue()), takeUntilDestroyed()).subscribe(() => {
       this.formChange.emit(this.buildFormValue());
     });
-  }
-
-  public onOrgSearch(event: AutoCompleteCompleteEvent): void {
-    this.orgSearchQuery.set(event.query.trim());
-  }
-
-  public onOrgSelect(event: AutoCompleteSelectEvent): void {
-    const org = event.value as OrgSearchResult;
-    this.form.patchValue({ company: org.name, organizationID: org.id });
-  }
-
-  public onOrgClear(): void {
-    this.form.patchValue({ company: '', organizationID: '' });
-    this.orgSearchQuery.set('');
-  }
-
-  public onOrgBlur(): void {
-    const orgValue = this.orgForm.get('company')?.value;
-    if (!orgValue) {
-      this.form.patchValue({ company: '', organizationID: '' });
-      this.form.get('company')?.markAsTouched();
-    }
-  }
-
-  private initOrgSuggestions(): Signal<OrgSearchResult[]> {
-    return toSignal(
-      toObservable(this.orgSearchQuery).pipe(
-        debounceTime(400),
-        switchMap((query) => {
-          if (!query) {
-            return of([]);
-          }
-          return this.eventsService.searchOrganizations(query).pipe(
-            map((response) => response.data),
-            catchError(() => {
-              this.messageService.add({
-                severity: 'error',
-                summary: 'Error',
-                detail: 'Failed to search organizations. Please try again.',
-              });
-              return of([]);
-            })
-          );
-        })
-      ),
-      { initialValue: [] }
-    );
   }
 
   private buildFormValue(): VisaRequestApplicantInfo {
