+++ packages/shared/src/interfaces/events.interface.ts
@@ -437,6 +437,7 @@ export interface TravelFundAboutMe {
   citizenshipCountry: string;
   profileLink: string;
   company: string;
+  organizationID?: string;
   canReceiveFunds: string;
   travelFromCountry: string;
   openSourceInvolvement: string;
